Boot loader - code cleanup.
