+++ boot/bootutil/test/src/testcases/boot_test_invalid_hash.c
@@ -60,7 +60,7 @@ TEST_CASE(boot_test_invalid_hash)
       &tlv, sizeof(tlv));
     TEST_ASSERT(rc == 0);
 
-    rc = boot_set_pending(1);
+    rc = boot_set_pending();
     TEST_ASSERT(rc == 0);
 
     boot_test_util_verify_all(&req, BOOT_SWAP_TYPE_NONE, &hdr0, NULL);
