+++ boot/bootutil/test/src/testcases/boot_test_nv_bs_11_2areas.c
@@ -54,7 +54,7 @@ TEST_CASE(boot_test_nv_bs_11_2areas)
     boot_test_util_write_hash(&hdr0, 0);
     boot_test_util_write_image(&hdr1, 1);
     boot_test_util_write_hash(&hdr1, 1);
-    rc = boot_set_pending(1);
+    rc = boot_set_pending();
     TEST_ASSERT_FATAL(rc == 0);
 
     boot_test_util_swap_areas(2, 5);
