+++ mgmt/imgmgr/include/imgmgr/imgmgr.h
@@ -66,6 +66,7 @@ int imgr_my_version(struct image_version *ver);
 
 uint8_t imgmgr_state_flags(int query_slot);
 int imgmgr_state_slot_in_use(int slot);
+int imgmgr_state_test_slot(int slot);
 
 #ifdef __cplusplus
 }
